feat: add generic next middleware
