+++ .changeset/brave-steaks-fly.md
@@ -1,0 +1,5 @@
+---
+"@cipherstash/nextjs": minor
+---
+
+Implemented a generic Next.js jseql middleware.
